requirements.txt
Flask==3.0.3
flask_cors==4.0.1
mysqlclient==2.2.4
mysql-connector-python==8.4.0
jwt==1.3.1